Make FixedLifespanScheduler a top level class
